chore: extract rendering functions to a new file
